txt_search_image/requirements-local.txt
# Local 模式依赖 (本地加载 CLIP 模型)
# 安装: pip install -r requirements-local.txt
#
# 注意: PyTorch 建议根据你的 CUDA 版本单独安装:
#   CPU: pip install torch torchvision
#   CUDA 11.8: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#   CUDA 12.1: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

# Milvus 向量数据库
pymilvus>=2.3.0

# 基础依赖
numpy>=1.21.0
Pillow>=9.0.0
rich>=12.0.0

# PyTorch (建议单独安装适合你 CUDA 版本的)
torch>=1.13.0
torchvision>=0.14.0

# clip-as-service 服务端 (包含模型加载功能)
clip-as-service[server]

# 中文 CLIP 支持 (可选)
# cn_clip
